Loading bar added
